requirements.txt
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-filters==0.2.1
djangorestframework==3.9.2
djangorestframework-jsonapi==2.7.0
inflection==0.3.1
psycopg2==2.8.1
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
